@@ -1,109 +1,63 @@
 {
   "name": "portfolio",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "tsc",
+    "preview": "vite preview"
+  },
   "dependencies": {
     "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.9.3",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.15",
-    "@headlessui/react": "^1.4.2",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@mui/material": "^5.0.0",
-    "@progress/kendo-licensing": "^1.2.0",
-    "@progress/kendo-react-animation": "^4.9.0",
-    "@progress/kendo-react-buttons": "^4.9.0",
     "@react-spring/parallax": "^9.2.4",
     "@reduxjs/toolkit": "^1.6.1",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@themesberg/react-bootstrap": "^1.4.1",
+    "@types/react-router-dom": "^5.3.3",
     "animate.css": "^4.1.1",
     "aos": "^3.0.0-beta.6",
     "axios": "^0.21.4",
-    "body-parser": "^1.19.0",
-    "bootstrap": "^5.1.1",
+    "bootstrap": "^5.3.2",
     "browserslist": "^4.16.5",
-    "caniuse-lite": "^1.0.30001365",
-    "express": "^4.17.1",
-    "framer-motion": "^4.1.17",
-    "glob-parent": "^5.1.2",
-    "gulp": "^4.0.2",
-    "mdbreact": "^5.1.0",
-    "mongodb": "^4.1.2",
-    "node-sass": "^6.0.1",
-    "npm": "^7.24.0",
+    "localforage": "^1.10.0",
+    "match-sorter": "^6.3.1",
+    "node-sass": "^9.0.0",
+    "npm": "^10.1.0",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-animate-on-scroll": "^2.1.5",
-    "react-awesome-loaders": "^0.1.37",
-    "react-bootstrap": "^2.0.0-beta.6",
-    "react-countup": "^5.2.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.0",
+    "react-countup": "^6.4.2",
     "react-dark-mode-toggle": "^0.2.0",
-    "react-dom": "^17.0.2",
-    "react-floating-button-menu": "^2.1.1",
-    "react-floating-dropdown": "^1.0.4",
-    "react-hover-animation": "^2.1.1",
-    "react-icons": "^4.4.0",
-    "react-image-overlay": "^1.0.5",
-    "react-load-animations": "^1.0.2",
-    "react-overlays": "^5.1.1",
-    "react-router-dom": "^5.3.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
+    "react-router-dom": "^6.16.0",
     "react-router-hash-link": "^2.4.3",
-    "react-scripts": "^4.0.3",
-    "react-scroll-motion": "^0.2.1",
-    "react-scroll-parallax": "^2.4.0",
-    "react-spring": "^9.5.0",
-    "react-switch": "^6.0.0",
-    "react-text-transition": "^1.3.0",
     "react-tilty": "^2.0.3",
-    "react-transition-group": "^4.4.2",
-    "react-typing-animation": "^1.6.2",
-    "reactjs-typing-effect": "^1.0.8",
-    "semantic-ui-css": "^2.4.1",
-    "styled-components": "^5.3.1",
+    "sass": "^1.68.0",
+    "sort-by": "^1.2.0",
+    "styled-components": "^5.3.11",
     "swiper": "^7.0.6",
-    "typeit-react": "^2.0.3",
     "web-vitals": "^1.0.1",
     "yargs-parser": "^20.2.9"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "client-i": "cd src && npm i"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "resolutions": {
-    "glob-parent": "^6.0.1"
-  },
   "devDependencies": {
-    "autoprefixer": "^10.3.4",
-    "css-loader": "^6.2.0",
-    "postcss": "^8.3.6",
-    "sass-loader": "^12.1.0",
-    "style-loader": "^3.2.1",
-    "tailwindcss": "^2.2.15"
+    "@types/aos": "^3.0.5",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@types/styled-components": "^5.1.28",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react-swc": "^3.3.2",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "postcss": "^8.4.31",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.5"
   }
 }
